Merge branch 'dev' of http://gitlab.voxteneo.co.id/.net/hms-fms into dev

# Conflicts:
#	.gitignore
#	UpgradeLog.htm

+++ FMS.BusinessObject/Dto/FuelOdometerDto.cs
@@ -8,7 +8,7 @@ namespace FMS.BusinessObject.Dto
 {
     public class FuelOdometerDto
     {
-        public int MstFuelOdometerId { get; set; }
+        public long MstFuelOdometerId { get; set; }
         public string VehicleType { get; set; }
         public string PoliceNumber { get; set; }
         public string EmployeeId { get; set; }
